退選邏輯修正

+++ unselect.php
@@ -51,56 +51,61 @@ if (isset($_POST['code'])) {
     $codeClass = array();
     $className = null;
     $noCourse = false;
+    $isMajor = false;
 
-    $sql = "SELECT sum(credit) as totalcredit , course_id  FROM select_list  NATURAL JOIN course where student_id = \"" . $studentID . "\";";
+
+    $sql = "SELECT  DISTINCT course_id ,credit FROM select_list  NATURAL JOIN course where student_id = \"" . $studentID . "\";";
     $result = mysqli_query($conn, $sql) or die('MySQL query error');
     if ($result->num_rows > 0) {
         while ($row = mysqli_fetch_array($result)) {
             array_push($mycourse, $row['course_id']);
-            $totalCredit = $row['totalcredit'];
+            $totalCredit =$totalCredit+$row['credit'];
         }
     }
-
-    $sql = "SELECT credit, class_name FROM course where course_id = \"" . $code . "%\";";
+    $sql = "SELECT credit, class_name ,major FROM course where course_id = \"" . $code . "%\";";
     $result = mysqli_query($conn, $sql) or die('MySQL query error');
     if ($result->num_rows > 0) {
         while ($row = mysqli_fetch_array($result)) {
-            array_push($codeClass, $row['class_name']);
+            array_push($codeClass, [$row['class_name'],$row['major']]);
             $codeCredit = $row['credit'];
         }
+
+        print_r($codeClass);
     } else {
         $noCourse = true;
         echo '<p>查無此課程<p>';
     }
 
     if ($noCourse != true) {
-
         if (!in_array($code, $mycourse)) {
             echo '你沒有選此課程';
             $noJoin = true;
         }
-        if ($totalCredit - $codeCredit < 9 && $codeCredit > 0 && $totalCredit > 0) {
-            $underCredit = true;
-            echo '<p>退選後學分低於9學分<p>';
-        }
+        if ($noJoin != true) {
+            if ($totalCredit - $codeCredit < 9 && $codeCredit > 0 && $totalCredit > 0) {
+                $underCredit = true;
+                echo '<p>退選後學分低於9學分<p>';
+            }
 
-        if ($underCredit != true && $noJoin != true) {
-            $sql = "SELECT class_name FROM student  where student_id = \"" . $studentID . "\";";
-            $result = mysqli_query($conn, $sql) or die('MySQL query error');
+            if ($underCredit != true && $noJoin != true) {
+                $sql = "SELECT class_name FROM student  where student_id = \"" . $studentID . "\";";
+                $result = mysqli_query($conn, $sql) or die('MySQL query error');
 
-            if ($result->num_rows > 0) {
-                while ($row = mysqli_fetch_array($result)) {
-                    $className = $row['class_name'];
-                }
-
-                for ($i = 0; $i < count($codeClass); $i++) {
-                    if ($className == $codeClass[$i]) {
-                        echo '退選此課程是必修課';
+                if ($result->num_rows > 0) {
+                    while ($row = mysqli_fetch_array($result)) {
+                        $className = $row['class_name'];
+                    }
+                    for ($i = 0; $i < count($codeClass); $i++) {
+                        if ($className == $codeClass[$i][0]) {
+                            echo '<p>退選此課程是必修課<p>';
+                        }
                     }
                     $sql = "DELETE  FROM select_list  where course_id = \"" . $code . "\" AND student_id = \"" . $studentID . "\" ;";
                     $result = mysqli_query($conn, $sql) or die('MySQL query error');
                     echo '退選成功';
+
                 }
+
             }
 
         }
